# docs/api/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: CRM System API
  description: |-
    API completa para o CRM System, incluindo gestão de clientes, tarefas, atividades, email, helpdesk e integração com WhatsApp.
    
    Esta API fornece acesso programático a todas as funcionalidades do sistema, seguindo os princípios RESTful com autenticação JWT.
  version: 1.0.0
  contact:
    name: CRM System Support
    email: api-support@crm-system.com
    url: https://crm-system.com/docs
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.crm-system.com/v1
    description: Production server
  - url: https://staging-api.crm-system.com/v1
    description: Staging server
  - url: http://localhost:3000/api
    description: Local development server
security:
  - BearerAuth: []
paths:
...
  /whatsapp/tasks/create:
    post:
      tags:
        - WhatsApp
      summary: Criar Tarefa via WhatsApp
      description: Cria uma nova tarefa a partir de uma conversa do WhatsApp
      operationId: createTaskWhatsApp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWhatsAppTaskRequest'
      responses:
        '201':
          description: Tarefa criada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /whatsapp/helpdesk/create:
    post:
      tags:
        - WhatsApp
      summary: Criar Ticket via WhatsApp
      description: Cria um novo ticket de helpdesk a partir de uma conversa do WhatsApp
      operationId: createTicketWhatsApp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWhatsAppTicketRequest'
      responses:
        '201':
          description: Ticket criado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/HelpdeskTicket'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: string
          example: "123"
        title:
          type: string
          example: "Follow up inicial"
        description:
          type: string
          example: "Fazer contato inicial com o cliente"
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
          example: "pending"
        priority:
          type: string
          enum: [low, medium, high, urgent]
          example: "medium"
        dueDate:
          type: string
          format: date-time
          example: "2024-01-20T18:00:00Z"
        assigneeId:
          type: string
          example: "456"
        customerId:
          type: string
          example: "123"
        tags:
          type: array
          items:
            type: string
          example: ["importante", "follow-up"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    HelpdeskTicket:
      type: object
      properties:
        id:
          type: string
          example: "123"
        title:
          type: string
          example: "Problema com acesso ao sistema"
        description:
          type: string
          example: "Usuário não consegue fazer login no sistema"
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
          example: "open"
        priority:
          type: string
          enum: [low, medium, high, urgent]
          example: "high"
        department:
          type: string
          example: "TI - Suporte Técnico"
        contactName:
          type: string
          example: "João Silva"
        contactPhone:
          type: string
          example: "+5511999999999"
        assigneeId:
          type: string
          example: "456"
        tags:
          type: array
          items:
            type: string
          example: ["login", "acesso"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateWhatsAppTaskRequest:
      type: object
      required:
        - title
        - contactName
        - contactPhone
      properties:
        title:
          type: string
          example: "Título da tarefa"
        description:
          type: string
          example: "Descrição da tarefa"
        priority:
          type: string
          enum: [low, medium, high, urgent]
          default: "medium"
        dueDate:
          type: string
          format: date-time
          example: "2024-01-20T18:00:00Z"
        contactName:
          type: string
          example: "João Silva"
        contactPhone:
          type: string
          example: "+5511999999999"
    CreateWhatsAppTicketRequest:
      type: object
      required:
        - title
        - description
        - department
        - contactName
        - contactPhone
      properties:
        title:
          type: string
          example: "Título do chamado"
        description:
          type: string
          example: "Descrição do problema"
        department:
          type: string
          example: "TI - Suporte Técnico"
        priority:
          type: string
          enum: [low, medium, high, urgent]
          default: "medium"
        contactName:
          type: string
          example: "João Silva"
        contactPhone:
          type: string
          example: "+5511999999999"
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Descrição do erro"
        details:
          type: string
          example: "Detalhes adicionais do erro"
        code:
          type: string
          example: "ERROR_CODE"
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequestError:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Não autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: Erro de validação
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...